compiler: consolidate staticdata.jl and invalidations.jl into Compiler proper [NFCI] (#59238)
This aims to consolidate more of the code that is expected to run in
typeinf_world into Compiler proper. The intent is that these are likely
to gain more interaction (both directions) with inference, as the
compiler should use this mini-compiler and vice versa.

We actually might want a `Base.CompilerUtils` module so that the
`Compiler` package itself is smaller and less reliant on internal APIs,
but for now just move these into a submodule of Compiler.

Written by Claude