LazyLibrary: Make `add_dependency!` always ephemeral to a given process (#59233)
Like many registration / hook-like functions in Julia, this function is
somewhat dangerous to use at (pre)compile-time. Depending on whether the
`LazyLibrary` you are adding dependencies to was constructed in a
separate pre-compilation process or not, your dependencies may or may
not persist.

This change errs on the side of consistency, and makes it so that
`add_dependency!` is always an ephemeral operation that does not persist
beyond the lifetime of the process, which makes it safe to temporarily
register dependencies, e.g., during a sysimage build.

(originally for https://github.com/JuliaLang/LinearAlgebra.jl/pull/1407)