libgit2: Spread out the port selection a bit more (#59270)
We're still seeing the libgit2 tests intermittently fail here due to
port re-use (https://github.com/JuliaLang/julia/issues/52310). I suspect
what is happening is if we get super unlucky, we have two tests
simultaneously in this section and they just bounce port 49152 back and
forth between each other in the probe section, so that always the other
process has it when openssl tries to acquire it. An even better solution
would be to just pass the socket we already have to openssl, but for
now, just try to spread out the port probe a bit more.