Remove throw for iterator of unknown size in last(::Zip)
Although it may be useful to through an ArgumentError rather than a
MethodError if we can't get the length of an iterator whose size is
unknown, SizeUnknown does not guarantee that length will throw.
Therefore, we remove this check and defer to a MethodError if the
iterator does not implement length.

Suggestion by @LilithHafner