Fix JET-reported issues in `BinaryPlatforms.triplet(::AbstractPlatform)` (#59343)
The issue is that JET reported that in `libgfortran_version(p).major`
the first arg of `getproperty` could be `nothing`. This is already
checked in the previous line, but in a way that the compiler cannot
remember until that call. Putting it into a variable should fix that.

It would be great if this could get backported to at least 1.12 (1.10
and 1.11 would also be great), since that is where people try to use JET
for their packages, and this reduces the Base noise in the output.