begin to encapsulate global state variables
Combine all exported (and some of the internal) global variables into a
single (or rather a couple) explicit structs. The eventual goal is to
support sub-runtime images, where it can be possible to load a trimmed
image into the same process, and be able to distinguish the exported
state for each without conflict.

The static-analysis is run with variables like before since it is
easier for the analysis to see them as separate.