bpart: Fix reresolution logic on export value changes
Fixes #59272. This code was originally introduced in
e7efe42ece419c04d170887d856609286e63836b. The design
of the binding partitions underwent several changes,
so I'm not fully sure if it was correct at the time,
but regardless, it was rendered incorrect by
60a9f6992d82789e23a1e71f1e3c7402b1483b4f.
In the new design, even a change to the value of a
binding (not just its visibility) can affect the
resolution outcome, so a full re-resolution is always
required.

Fix identified by Claude (in one of several rollouts).
Correct fix among them identified by me.
Actual change/test by me.