Add lazily-computed-default ScopedValue variant
Test has `Test.record_passes()` which gets `OncePerProcess` initialized
to the value of `ENV["JULIA_TEST_RECORD_PASSES"]`. However, in another
palce the test suite once to override this value for its scope, and does
this by setting the environment variable, relying on OncePerProcess
picking this up. However, this is not at all semantically guaranteed
and of course fails if anybody happened to have a testset before
(which is possible, because this is used by Base.runtests). It would
be much better if we could use a ScopedValue for this like the other
`Test` setting state (after #53462). However, there currently isn't
any way to once-initialize the ScopedValue default. This PR aims to
fix that by letting the scoped value default be an evaluated callback.