Properly handle (re)escaping and replace_ref_begin_end! interactions
This commit refactors some of the recent additions to handle reescaping
in InteractiveUtils, along with tests to cover various hygiene contexts.

`replace_ref_begin_end!` was recently replaced in InteractiveUtils
by its low-level counterpart, `replace_ref_begin_end_!`, to better
handle escapes. However, we were using it in a way that
required duplication of the code generation logic to match the code
that we emit it for other code paths. Because the logic is now more complex with
the recently added `use_signature_tuple` behavior, it is preferable to
allow reuse of the same code path for all code we emit. This commit
implements this refactor, which unfortunately also requires a manual
escape of the temporary variable that is emitted by
`replace_ref_begin_end!` in its current design.