atomics: correctly implement padding write of 12 byte atomics with 4 byte pointer
On 32 bit machines, for an atomic of size 9 to 12 bytes, the result fits
in the 16 byte pool, but only with a maximum write of 12 bytes. This was
accidentally doing a 16 byte write instead, which could smash the type
tag field (usually will NULL) of the next object. Not sure how to test,
since just noticed this while reading the code.