inference: reinfer missing code for inlining
When code is potentially needed for inlining, but missing, be sure to
regenerate it during inference with the correct `ci_meets_requirement`
flags (SOURCE_MODE_GET_SOURCE instead of NOT_REQUIRED).

Also be sure to use CACHE_MODE_LOCAL, since it needs to actually end up
in the local cache (as we know the global cache already has it, but
volatile may discard it before inlining).