Remove obsolete paragraph on struct redefinition from FAQ (#59399)
The FAQ for Julia v1.12 and master still contains a paragraph that
explains how to work around the impossibility to redefine a type in a
REPL session. This paragraph is now obsolete thanks to #57253 so I
propose to simply remove it.

(cherry picked from commit 0c1fab4fd2a3943801bab055d45bba2a40fc1080)