decrease method count for `eltype` for `Tuple`, decrease `max_methods` (#58788)
The `eltype` function was one of the few functions in the sysimage with
a `max_methods` value (the world-splitting threshold) greater than the
default. This was a workaround for the unnecessarily large number of
methods of `eltype(::Type{<:Tuple})`. The `max_methods` value was
increased in PR #48322 to help effect inference.