generic `size`: avoid method static parameters and abstract type assert
The generic method `size(::AbstractArray, ::Any)` currently has:

* method static parameters

* an abstract type assert (`::Integer`) in the method body

PR #59442 aims to make this generic method be used for `Array` and
`Memory`, by deleting the more specific methods.

It is my understanding that method static parameters and abstract type
asserts can cause performance issues, when they're not optimized out,
which happens when the argument types are not concretely inferred.

So I'm putting up this PR, which eliminates the method static
parameters and the `typeassert`, to prevent any possible regressions
from PR #59442.

Another thing that is necessary for preserving good abstract inference
is to convert the index to `Int` before using it. This is correct
it there can't be more than `typemax(Int)` dimensions anyway.