reducedim: force specialization of some dims arguments
In Julia, the colon (`:`, `Colon`) is a function for constructing
ranges. We have:

```julia
(:) isa Function
```

However, sadly, as an instance of punning, the same object is often
also used as an index for choosing an entire dimension.

A problem is that `Function` objects are a special case when it comes
to specializing method code for given argument types: by default a `:`
argument will get specialized as a `Function`, instead of as a `Colon`.

This change forces specialization of dims arguments in a bunch of
methods in `base/reducedim.jl`. Probably similar changes would be
beneficial across the code base, and across the ecosystem.

I believe this change will make the sysimage more resistant to
invalidation, we'll see once it builds.