spawn: Add support for setuid/setgid (#59420)
Libuv support changing the uid/gid of the spawned process. This is a bit
niche but is useful if your process is running privileged in a sandbox
and wants to drop privilege for certain operations on user code. Since
libuv supports it, let's wire up the necessary bits for it on the julia
side. Written by Claude.