recache all previously cached / new codeinst
Regardless of existing cache content, these may be useful to be able to
examine or reuse in the future. Even if it is just to JIT new code
later, we already have this object in memory, so it costs nothing to be
able to enumerate them from this linked list. To accomplish that there
are four changes to cache insertion:
 - sorting: add a new item to the cache will insert at the optimal
   location for runtime performance
 - idempotency: add a new item to the cache may move it to a new better
   place, but won't break the cache
 - universal inclusion: even when the method has no recorded edges, it
   still might require scanning for Bindings or coverage
 - correct initialization: edges needs to be set, even if uninferred (so
   just empty svec), on everything in cache, since many places examine
   the cache expect to follow those