objconv: Update to version 2.55 (#59491)
This also updates the build recipe to download objconv from the same git
repository that we use to build `Objconv_jll`. This is a good idea
because upstream does not provide specific versions of the objconv
library, it only provides the current version, and that makes Julia
builds non-reproducible.

(For example, the objconv version number was listed as 2.53, but
upstream provided 2.54, and we had no good way to check.)