begin to encapsulate global state variables (#59400)
Combine all exported (and some of the internal) global variables into a
single (or rather a couple) explicit structs. The eventual goal is to
support sub-runtime images, where it can be possible to load a trimmed
image into the same process, and be able to distinguish the exported
state for each without conflict. This also helps remove a lot of code
duplication for defining these, since we can use the XX style to copy
them in a group, instead of listing them explicitly.