document that unsafe_string accepts Cstring (#56907)
Document that `unsafe_string` allows its argument to be a `Cstring`.

There is also another method that accepts `Cwstring`, which transcodes
from a NUL-terminated `wchar_t` array (assumed to be either UTF-32 or
UTF-16 depending on the width of `wchar_t`). Should that be documented
in the same docstring or in a different docstring?

---------

Co-authored-by: Max Horn <max@quendi.de>
Co-authored-by: Viral B. Shah <viral@juliacomputing.com>