docs: Add headers for easier Core, Base and stdlib reference (#55311)
This adds one new section (h2) and a nested subsection (h3) to the
landing page of the Julia documentation website and associated PDF file.
The new section provides a quick overview of the differences between
`Core`, `Base` and the standard library, as well as a list of links to
standard library packages.

The list of standard library packages is thus duplicated in the website
sidebar, but I don't think this is an issue because the sidebar is
primarily intended for navigation, whereas this list is primarily
intended to be used for referencing purposes.

Partially addresses #28712, although I don't provide short descriptions
of each standard library package. I think the appropriate place for that
would be in module-level docstrings of the packages themselves, which
are not currently available.

---------

Co-authored-by: Viral B. Shah <viral@juliacomputing.com>
(cherry picked from commit abb104c56f5a43e6511bb5cfa1a6586a4fc8cd46)