codegen: mark svec fields GC as release stores
Without a release store, it seems LLVM considers it a data race to have
read the initial state on another thread. Marking this as a release
store seems sufficient to prevent that optimization. It is also more
consistent with how we initialize and write to most other structs,
particularly since #55767.

Fixes #59547 (more)