Interpret `module` expressions at top level in `eval()`
This is a step toward an iteration interface for lowering which can
return a sequence of CodeInfo to be evaluated for top level and module
expressions.

This also restricts lowering of module expressions to be syntactically
at top level (ie, not inside a top level thunk), consistent with the
existing way that they're handled in eval.