InteractiveUtils: Support unannotated values after `Vararg` type annotation (#59594)
Implements #58900, ignoring any arguments coming after a
`::Vararg`/`::Vararg{...}`/`::T...` annotation, after having ensured
that no explicit type annotations were present on the dropped arguments.

The logic to detect whether it is a call to `Vararg` is a bit fishy, as
we assume `A.B.C.Vararg` is the same as `Core.Vararg`, and we ignore
that hygiene may apply another scope, but I think it is a reasonable
behavior/expectation that if someone annotates an argument with`Vararg`
they mean `Core.Vararg`, not a custom `Vararg` type. I don't feel
strongly about that though.

---------

Co-authored-by: Cody Tapscott <84105208+topolarity@users.noreply.github.com>