Remove GPL libraries from the Julia build for binary-dist target (#59627)
Currently we support removing GPL dependencies in the full source build.
This will also remove the GPL dependencies from the binary-dist target
when built with JLLs.

I almost feel like it would be simpler to have a new
SuiteSparse_NOGPL_jll package. Then in the default build, things stay as
they are. In the no gpl build use the new JLL. In the no GPL build, if
someone then tries to use a GPL SuiteSparse library, a warning can be
printed asking them to get a different build of Julia.

@DilumAluthge @andreasnoack @giordano Thoughts?

Co-authored-by: Viral B. Shah <ViralBShah@users.noreply.github.com>