Apply an inference barrier to annot printstyled (#59596)
In 6d78a4a we introduced an inference barrier to the various
AnnotatedX/StyledStrings methods whose implementations are frustratingly
split between Base and a Stdlib, to ease the pain of invalidations.

As has recently been pointed out in the re-opened JuliaLang#57997, this missed
a `printstyled` method. We address that here, together with a tweak to
the StyledStrings library.

Reported-by: Neven Sajko <s@purelymail.com>