Base: Add `count` argument to `Base.summarysize` (#58874)
In many cases, number of allocations is more important than memory
allocated (w.r.t. performance).

This adds a small utility to `summarysize` to allow counting the number
of unique reachable objects, which can be useful to understand, e.g., if
you are performing many small allocations or how many of the allocations
in a function ended up reachable from the result.

---------

Co-authored-by: Jameson Nash <vtjnash@gmail.com>