remove Base.memhash global
This became unsound to use even though it was preserved "to avoid
breakage" in v1.13, since continuing to use it would give incorrect hash
results, which could result in corrupt dictionaries.

Since #59691, these broken `hash` methods can now simply be deleted as
they no longer provide any value.

It is hard to say whether this is technically breaking or not as a
change. It causes packages to go from giving subtly wrong answers (the
worst kind of wrong) to crashing in v1.13, until the offending incorrect
methods are deleted.