lowering: increment world age after toplevel expressions
When a macro expands to `:toplevel` expression, it seems reasonable
(as would be if the expression were not wrapped in `:toplevel`) that
any contained struct definitions be available thereafter. For example,
this applies to `@enum`, which countrly causes world age erros when
attempting to define a method on an enum within the same expression.
Fix this by having lowering insert and explicit `latestworld` after
toplevel.

Fixes #59429