[Mmap] Turn `PAGESIZE` into a function
`PAGESIZE` is currently fixed at compile time, but the machine where Julia runs
may have a different memory page size than the one where the stdlib was compiled
on.  It's not worth making this a `OncePerProcess` object because the underlying
function call is very cheap, cheaper than the overhead of `OncePerProcess`.