fix performance of ProductIterator for non-concrete iterators (#59711)
This uses inferred element types to make the iterator state type-stable,
enabling inlining, which can enable the runtime to see the real types.
This is essentially the transform that we want inference to infer and
codegen to implement, except made explicit so that we can rely on it
actually happening.

Fix #56607

```
julia> @btime f()
  574.500 ns (2 allocations: 272 bytes) # non-const op
  253.878 ns (2 allocations: 272 bytes) # const op
```