Extend backdating checks for imports and globals
Previously, backdated binding warnings were only emitted for constants.
This commit extends the backdating mechanism to handle imports and globals.
This change was requested by triage to improve compatibility of the new
binding partition mechanism given issues like #58511.

With this change the behavior is as follows:
```julia
julia> function foo()
           include_string(Main, "a = 42")
           return a
       end

julia> foo()
WARNING: Detected access to binding `Main.a` in a world prior to its definition world.
  Julia 1.12 has introduced more strict world age semantics for global bindings.
  !!! This code may malfunction under Revise.
  !!! This code will error in future versions of Julia.
Hint: Add an appropriate `invokelatest` around the access to this binding.
To make this warning an error, and hence obtain a stack trace, use `julia --depwarn=error`.
42
```

With `julia --depwarn=error`, the access is rejected:

```julia
julia> foo()
ERROR: UndefVarError: `a` not defined in `Main`
```

I am still not a huge fan of the backdating mechanism in general, but
I think this is the least objectionable of the alternatives and provides
a big warning to let people know that something is wrong.