docs: manual: fix inclusion and location of `:` doc strings (#59262)
This fixes issues with the location and lack of inclusion of the docs
strings of `:` into the Manual:

* The doc string for the metaprogramming functionality was mistakenly
included into the page on math, instead of into Essentials.

* The other doc strings, for the range functionality, were missing
entirely from the manual, so include them into the page on math.

These doc bugs were introduced all the way back in PR #19609!