remove Base.memhash global (#59697)
This became unsound to use even though it was preserved "to avoid
breakage" in v1.13, since continuing to use it would give incorrect hash
results, which could result in corrupt dictionaries and unsound
programs.

Since #59691, these broken `hash` methods in the ecosystem can now
simply be deleted as they no longer provide any value.

It is hard to say whether this is technically breaking or not as a
change. It causes packages to go from giving subtly wrong answers (the
worst kind of wrong) to crashing in v1.13, until the offending incorrect
methods are deleted.

n.b. this is expected to break several packages (notably among them,
JuliaInterpreter, which has several tests specifically just for the
existence of this global), which will require some ecosystem updates.
The update should just be to delete the offending method (they is now
redundant, undesirable, and have been giving unsafely buggy answers),
but we need to do some work to identify those places and release new
versions.